WIP extract similarity functions
