+++ project/src/main/java/es/ubu/lsi/avrela/apm/model/IssueSimilarity.java
@@ -1,8 +1,9 @@
 package es.ubu.lsi.avrela.apm.model;
 
+import es.ubu.lsi.avrela.similarity.Jaccard;
+import es.ubu.lsi.avrela.similarity.JaroWinklerDistance;
 import java.util.EnumMap;
 import java.util.HashSet;
-import java.util.Set;
 
 public class IssueSimilarity {
 
@@ -13,17 +14,7 @@ public class IssueSimilarity {
   }
 
   public static double calculate(Issue a, Issue b, EnumMap<Feature, Double> featureWeights) {
-    // Calculate the intersection and union between the two sets of labels
-    Set<String> labelIntersection = new HashSet<>(a.getLabels());
-    labelIntersection.retainAll(b.getLabels());
-    int labelIntersectionSize = labelIntersection.size();
-    Set<String> labelUnion = new HashSet<>(a.getLabels());
-    labelUnion.addAll(b.getLabels());
-    int labelUnionSize = labelUnion.size();
-
-    // Calculate the Jaccard similarity using the formula:
-    // (size of intersection) / (size of union)
-    double labelSimilarity = (double) labelIntersectionSize / labelUnionSize;
+    double labelSimilarity = Jaccard.similarityOf(new HashSet<>(a.getLabels()), new HashSet<>(b.getLabels()));
 
     int stateSimilarity = ( a.getState() == b.getState()) ? 1 : 0;
