+++ project/src/test/java/es/ubu/lsi/avrela/scm/model/CommitSimilarityTest.java
@@ -1,27 +1,11 @@
 package es.ubu.lsi.avrela.scm.model;
 
-import es.ubu.lsi.avrela.similarity.Jaccard;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 
 public class CommitSimilarityTest {
 
-  @Test
- public void jaccardSimilarityTest(){
-   Assertions.assertEquals(0.2, Jaccard.similarityOf( new HashSet<>(commitFiles12345()),
-       new HashSet<>(commitFiles1())));
-   Assertions.assertEquals(0.8, Jaccard.similarityOf( new HashSet<>(commitFiles1234()),
-       new HashSet<>(commitFiles12345())));
-   Assertions.assertEquals(0.0, Jaccard.similarityOf( null,
-       new HashSet<>(commitFiles12345())));
-   Assertions.assertEquals(0.0, Jaccard.similarityOf( new HashSet<>(commitFiles9()),
-       new HashSet<>(commitFiles12345())));
-   Assertions.assertEquals(1.0, Jaccard.similarityOf( new HashSet<>(commitFiles12345()),
-       new HashSet<>(commitFiles12345())));
- }
+
 
  public void commitSimilarity () {
    Commit commitA = Commit.builder()
